src/requirements.txt
Flask==1.0.3
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp37-cp37m-linux_x86_64.whl
torchvision==0.2.2
Werkzeug==0.15.4
Pillow==7.1.2
gunicorn==20.0.4

